fixes for feature 15
