+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_executable(Dictionary1 main1.c gestmemory.c gestbib.c dictionary.c metadata.c input.c)
-add_executable(Dictionary2 main2.c gestmemory.c gestrech.c dictionary.c metadata.c input.c)
-add_executable(Dictionary3 main3.c gestmemory.c gestorth.c dictionary.c metadata.c input.c)
-add_executable(Dictionary4 main4.c gestmemory.c gestexpr.c dictionary.c metadata.c input.c)
+add_executable(Dictionary1 main1.c gestmemory.c gestbib.c dictionary.c metadata.c input.c utils.c)
+add_executable(Dictionary2 main2.c gestmemory.c gestrech.c dictionary.c metadata.c input.c utils.c)
+add_executable(Dictionary3 main3.c gestmemory.c gestorth.c dictionary.c metadata.c input.c utils.c)
+add_executable(Dictionary4 main4.c gestmemory.c gestexpr.c dictionary.c metadata.c input.c utils.c)
 
 #Adding the md5 lib
 #target_link_libraries(DictionaryX md5)
